test_data_dummy/requirements.txt
paho-mqtt
python-dotenv
pytz